refactor(docker): use multistage and remove fonoster/base dep
With this approach the image is self-contained, the final size is reduced, and the runner (used to deploy) has a lower attack factor, as it has less available packages.

This approach also improves the use of docker cache, and makes the build time faster.

BREAKING CHANGE: this approach uses Yarn instead of NPM for improved package security and better conflict handling.